Add casts to get rid of warnings, add documentation

+++ box2dhandler.cpp
@@ -99,10 +99,14 @@ void box2dhandler::userGrab(int xPos, int yPos)
     for(b2Body* body : bodies)
     {
         //again uses the assumption that the mass is equal to the size
-        int size = body->GetMass();
+        //Note: this size cast is okay becuase only int sizes are made
+        int size = static_cast<int>(body->GetMass());
         b2Vec2 bodyPosition = body->GetPosition();
-        int xDif = abs(xPos - bodyPosition.x);
-        int yDif = abs(yPos - bodyPosition.y);
+
+        //Note: these cast sare okay, just rounding to the nearest pixel
+        int xDif = abs(xPos - static_cast<int>(bodyPosition.x));
+        int yDif = abs(yPos - static_cast<int>(bodyPosition.y));
+
         if(xDif < (size / 2) &&  yDif < (size / 2))
         {
             //can grab
@@ -130,6 +134,7 @@ void box2dhandler::userMove(int xPos, int yPos)
 
     if(yPos < 0)
         yPos = 0;
+
 
     grabbed->SetTransform(b2Vec2(xPos,yPos),grabbed->GetAngle());
 }
